requirements.txt
wandb
tqdm
numpy==1.26.4
matplotlib
torchio
torchinfo
torchmetrics
prodigyopt
monai
einops
h5py
hdf5plugin
PySide2
KDEpy
scikit-image
pytest
ptflops
bm4d